+++ frontend/src/utils/statusBadgeColor.ts
@@ -1,0 +1,9 @@
+export default function statusBadgeColor(status: string) {
+  if (status === 'Active') {
+    return 'success';
+  }
+  if (status === 'Seen recently') {
+    return 'warning';
+  }
+  return 'error';
+}
